pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bfsi_sales_agent"
version = "0.1.0"
description = "AI Sales Agent for BFSI sector"
readme = "README.md"
requires-python = ">=3.8.1"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "flask>=2.3.3",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "twilio>=8.5.0",
    "PyPDF2>=3.0.1",
    "loguru>=0.7.0",
    "elevenlabs>=0.2.26",
    "python-dateutil>=2.8.2",
    "tenacity>=8.2.3",
    "PyJWT>=2.8.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0"
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88